Miscellaneous changes

+++ app/graph_mood.py
@@ -91,9 +91,9 @@ def mood_figs(data, height = 500, width = 1200):
         y = float((v_u + v_l)/2)
         w = (a_u - a_l)
         h = (v_u - v_l)
-        alpha = float((1 + abs(5 - a_be)**4)*alpha_fudge) #***TODO FIX: Indicate better
+        alpha = int((1 + abs(5 - a_be)**4)*alpha_fudge) #***TODO FIX: Indicate better
 
-        c_val = float((1 + abs(5 - v_be)**4)*color_fudge)
+        c_val = int((1 + abs(5 - v_be)**4)*color_fudge)
 
         #Some bug in Bokeh or something; does not accept (R,G,B) tuples as colors
         if v_be >= 5:
